[svn-r27288] Bring revision #24622 from revise_chksum_retry branch to revise_chunks.  h5committested. #test-p fails but will be fixed with merges later: #26569 (revise_chksum_retry) + v3 metadata cache #27237 (trunk).

+++ test/accum_swmr_reader.c
@@ -65,7 +65,7 @@ main(void)
 	FAIL_STACK_ERROR;
 
     /* Verify the data read is correct */
-    if(HDmemcmp(buf, rbuf, 1024) != 0) 
+    if(HDmemcmp(buf, rbuf, (size_t)1024) != 0) 
 	TEST_ERROR;
 
     /* CLose the file */
